requirements.txt
appdirs
packaging
cryptography
pycryptodome
requests
